Fix node detail size issue
